Refactoring, adding configs
