+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=Example
-author=Nobody
+displayName=PK Build Indicator
+author=adamisafk
 support=
-description=An example greeter plugin
-tags=
-plugins=com.example.ExamplePlugin
+description=A RuneLite plugin which tries to read other player's skills and worn equipment to assume and label their PK account build.
+tags=pk,build,indicator,indicators,label,helper
+plugins=com.pkbuildindicator.PKBuildIndicatorPlugin
