+++ src/main/java/com/PKBuildIndicator/PKBuildIndicatorConfig.java
@@ -1,11 +1,11 @@
-package com.example;
+package com.PKBuildIndicator;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 
-@ConfigGroup("example")
-public interface ExampleConfig extends Config
+@ConfigGroup("PKBuildIndicator")
+public interface PKBuildIndicatorConfig extends Config
 {
 	@ConfigItem(
 		keyName = "greeting",
