+++ src/main/java/com/PKBuildIndicator/PKBuildIndicatorPlugin.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.PKBuildIndicator;
 
 import com.google.inject.Provides;
 import javax.inject.Inject;
@@ -14,15 +14,17 @@ import net.runelite.client.plugins.PluginDescriptor;
 
 @Slf4j
 @PluginDescriptor(
-	name = "Example"
+	name = "PK Build Indicators",
+	description = "Highlight players' account builds on-screen for PKing",
+	tags = {"builds", "highlight", "overlay", "players", "pking", "player", "killing" }
 )
-public class ExamplePlugin extends Plugin
+public class PKBuildIndicatorPlugin extends Plugin
 {
 	@Inject
 	private Client client;
 
 	@Inject
-	private ExampleConfig config;
+	private PKBuildIndicatorConfig config;
 
 	@Override
 	protected void startUp() throws Exception
@@ -46,8 +48,8 @@ public class ExamplePlugin extends Plugin
 	}
 
 	@Provides
-	ExampleConfig provideConfig(ConfigManager configManager)
+	PKBuildIndicatorConfig provideConfig(ConfigManager configManager)
 	{
-		return configManager.getConfig(ExampleConfig.class);
+		return configManager.getConfig(PKBuildIndicatorConfig.class);
 	}
 }
