+++ src/test/java/com/PKBuildIndicator/PKBuildIndicatorPluginTest.java
@@ -1,13 +1,13 @@
-package com.example;
+package com.PKBuildIndicator;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
 
-public class ExamplePluginTest
+public class PKBuildIndicatorPluginTest
 {
 	public static void main(String[] args) throws Exception
 	{
-		ExternalPluginManager.loadBuiltin(ExamplePlugin.class);
+		ExternalPluginManager.loadBuiltin(PKBuildIndicatorPlugin.class);
 		RuneLite.main(args);
 	}
 }
